+++ src/shared/openTelemetry.js
@@ -1,32 +1,53 @@
 "use strict";
 
+const _resilientRequire = (module) => {
+  try {
+    return require(module);
+  } catch {
+    // ignore
+  }
+};
+
 const cds = require("@sap/cds");
-let otel;
-try {
-  otel = require("@opentelemetry/api");
-} catch {
-  // ignore
-}
+const otel = _resilientRequire("@opentelemetry/api");
 
 const config = require("../config");
 
 const COMPONENT_NAME = "/shared/openTelemetry";
 
-const trace = async (context, label, fn, { attributes = {}, newRootSpan = false } = {}) => {
+const trace = async (context, label, fn, { attributes = {}, newRootSpan = false, traceContext } = {}) => {
   const tracerProvider = otel?.trace.getTracerProvider();
-  // Check if a real provider is registered
-  if (!config.enableCAPTelemetry || !tracerProvider || tracerProvider === otel.trace.NOOP_TRACER_PROVIDER) {
+  // TODO: extend check to validate if DT oneagent is available AND active
+  if (!config.enableTelemetry || !tracerProvider || tracerProvider === otel.trace.NOOP_TRACER_PROVIDER) {
     return fn();
   }
 
-  const tracer = otel.trace.getTracer("eventqueue");
-  const span = tracer.startSpan(`eventqueue-${label}`, {
-    kind: otel.SpanKind.INTERNAL,
-    root: newRootSpan,
-  });
+  const tracer = otel.trace.getTracer("@cap-js-community/event-queue");
+  const extractedContext = traceContext
+    ? otel.propagation.extract(otel.context.active(), traceContext)
+    : otel.context.active();
+  const span = tracer.startSpan(
+    `eventqueue-${label}`,
+    {
+      kind: otel.SpanKind.INTERNAL,
+      root: newRootSpan,
+    },
+    extractedContext
+  );
   _setAttributes(context, span, attributes);
-  const ctxWithSpan = otel.trace.setSpan(otel.context.active(), span);
+  const ctxWithSpan = otel.trace.setSpan(extractedContext, span);
+
+  return await _startOtelTrace(ctxWithSpan, traceContext, span, fn);
+};
+
+const _startOtelTrace = async (ctxWithSpan, traceContext, span, fn) => {
   return otel.context.with(ctxWithSpan, async () => {
+    if (traceContext) {
+      cds.log("/eventQueue/telemetry").info("Linked span:", span.spanContext());
+      const carrier = {};
+      otel.propagation.inject(ctxWithSpan, carrier);
+      cds.log("/eventQueue/telemetry").info("Extracted trace context by inject", carrier);
+    }
     const onSuccess = (res) => {
       span.setStatus({ code: otel.SpanStatusCode.OK });
       return res;
@@ -72,4 +93,13 @@ const _setAttributes = (context, span, attributes) => {
   }
 };
 
-module.exports = trace;
+const getCurrentTraceContext = () => {
+  if (!otel) {
+    return null;
+  }
+  const carrier = {};
+  otel.propagation.inject(otel.context.active(), carrier);
+  return carrier;
+};
+
+module.exports = { trace, getCurrentTraceContext };